+++ modules/client/CHANGELOG.md
@@ -24,6 +24,8 @@ TEMPLATE:
 - Underlying tokens queries in getProposal and getProposals
 - Plugins query failing due to prepared but not applied installations
 - Ipfs default endpoinst missing api/v0
+- Make the `network` parameter required on `getPluginInstallItem`
+
 ## [1.10.0-rc1]
 ### Fixes
 - Update common-client version