+++ modules/client/CHANGELOG.md
@@ -19,12 +19,12 @@ TEMPLATE:
 ## [UPCOMING]
 ### Fixes
 - Fix status calculation for token voting proposals
+- Make the `network` parameter required on `getPluginInstallItem`
 ## [1.10.1-rc1]
 ### Fixes
 - Underlying tokens queries in getProposal and getProposals
 - Plugins query failing due to prepared but not applied installations
 - Ipfs default endpoinst missing api/v0
-- Make the `network` parameter required on `getPluginInstallItem`
 
 ## [1.10.0-rc1]
 ### Fixes